+++ components/structure/pagination.tsx
@@ -1,0 +1,92 @@
+import React, { useState } from "react";
+
+export default function Pagination() {
+  const [currentPage, setCurrentPage] = useState(1);
+  const totalPages = 5;
+
+  const handlePageChange = (page: number) => {
+    setCurrentPage(page);
+  };
+
+  const renderPageNumbers = () => {
+    const pages = [];
+
+    for (let i = 1; i <= totalPages; i++) {
+      const isActive = i === currentPage;
+
+      pages.push(
+        <li key={i}>
+          <a
+            href="#"
+            className={`px-3 py-2 leading-tight ${
+              isActive
+                ? "text-blue-600 bg-blue-50 border-blue-300 hover:bg-blue-100 hover:text-blue-700 dark:border-gray-700 dark:bg-gray-700 dark:text-white"
+                : "text-gray-500 bg-white border-gray-300 hover:bg-gray-100 hover:text-gray-700 dark:bg-gray-800 dark:border-gray-700 dark:text-gray-400 dark:hover:bg-gray-700 dark:hover:text-white"
+            } ${isActive ? "pointer-events-none" : ""}`}
+            onClick={() => handlePageChange(i)}
+          >
+            {i}
+          </a>
+        </li>
+      );
+    }
+
+    return pages;
+  };
+
+  return (
+    <nav aria-label="Page navigation example">
+      <ul className="inline-flex items-center -space-x-px">
+        <li>
+          <a
+            href="#"
+            className={`block px-3 py-2 ml-0 leading-tight text-gray-500 bg-white border border-gray-300 rounded-l-lg hover:bg-gray-100 hover:text-gray-700 dark:bg-gray-800 dark:border-gray-700 dark:text-gray-400 dark:hover:bg-gray-700 dark:hover:text-white ${
+              currentPage === 1 ? "pointer-events-none" : ""
+            }`}
+            onClick={() => handlePageChange(currentPage - 1)}
+          >
+            <span className="sr-only">Previous</span>
+            <svg
+              aria-hidden="true"
+              className="w-5 h-5"
+              fill="currentColor"
+              viewBox="0 0 20 20"
+              xmlns="http://www.w3.org/2000/svg"
+            >
+              <path
+                fillRule="evenodd"
+                d="M12.707 5.293a1 1 0 010 1.414L9.414 10l3.293 3.293a1 1 0 01-1.414 1.414l-4-4a1 1 0 010-1.414l4-4a1 1 0 011.414 0z"
+                clipRule="evenodd"
+              />
+            </svg>
+          </a>
+        </li>
+        {renderPageNumbers()}
+        <li>
+          <a
+            href="#"
+            className={`block px-3 py-2 leading-tight text-gray-500 bg-white border border-gray-300 rounded-r-lg hover:bg-gray-100 hover:text-gray-700 dark:bg-gray-800 dark:border-gray-700 dark:text-gray-400 dark:hover:bg-gray-700 dark:hover:text-white ${
+              currentPage === totalPages ? "pointer-events-none" : ""
+            }`}
+            onClick={() => handlePageChange(currentPage + 1)}
+          >
+            <span className="sr-only">Next</span>
+            <svg
+              aria-hidden="true"
+              className="w-5 h-5"
+              fill="currentColor"
+              viewBox="0 0 20 20"
+              xmlns="http://www.w3.org/2000/svg"
+            >
+              <path
+                fillRule="evenodd"
+                d="M7.293 14.707a1 1 0 010-1.414L10.586 10 7.293 6.707a1 1 0 011.414-1.414l4 4a1 1 0 010 1.414l-4 4a1 1 0 01-1.414 0z"
+                clipRule="evenodd"
+              />
+            </svg>
+          </a>
+        </li>
+      </ul>
+    </nav>
+  );
+}
