--- pages/best.tsx
@@ -1,3 +1,0 @@
-export default function Best() {
-  return null;
-}
