+++ pages/menus/trend.tsx
