--- pages/new.tsx
@@ -1,3 +1,0 @@
-export default function New() {
-  return null;
-}
