--- pages/pdetails.tsx
@@ -1,3 +1,0 @@
-export default function Pdetails() {
-  return null;
-}
