+++ pages/prodDetails.tsx
@@ -1,0 +1,3 @@
+export default function ProDetails() {
+  return null;
+}
